requirements.txt
requests
websocket
tk
temp
alpaca-trade-api
pillow
plotly
ValeraLib
